Added Certificate-handling and some fixes

+++ D365BCOnAzureHelper/Export/Private/Disable-InternetExplorerESC.ps1
@@ -22,9 +22,10 @@ function Disable-InternetExplorerESC {
             $UserKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A8-37EF-4b3f-8CFC-4F3A74704073}"
             Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value 0 -Force
         }
-        
-        Write-Verbose "Stopping Process 'Explorer' to apply settings"
-        Stop-Process -Name Explorer -Force
+        if (Get-Process -Name Explorer -ErrorAction SilentlyContinue) {
+            Write-Verbose "Stopping Process 'Explorer' to apply settings"        
+            Stop-Process -Name Explorer -Force
+        }
         Write-Verbose "IE Enhanced Security Configuration (ESC) has been disabled."
     }
 }
